Rename `header.php` to `header_authenticated.php`
