+++ scripts/pi-hole/php/header_authenticated.php
